Arreglado todo lo técnico

+++ templates_c/e3a3b4e3bc4ee0d864ba57a8879d095d9ae5f2b7_0.file.libroId.tpl.php
@@ -1,75 +1,35 @@
 <?php
-/* Smarty version 4.2.1, created on 2023-10-15 07:43:29
+/* Smarty version 4.2.1, created on 2023-10-16 04:51:31
   from 'Z:\XAMPP\htdocs\WebII\TPE\templates\libroId.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_652b7c0118efc8_24745409',
+  'unifunc' => 'content_652ca5337e6dd8_99327200',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     'e3a3b4e3bc4ee0d864ba57a8879d095d9ae5f2b7' => 
     array (
       0 => 'Z:\\XAMPP\\htdocs\\WebII\\TPE\\templates\\libroId.tpl',
-      1 => 1697348572,
+      1 => 1697423937,
       2 => 'file',
     ),
   ),
   'includes' => 
   array (
     'file:header.tpl' => 1,
-    'file:error.tpl' => 1,
+    'file:formularioLibro.tpl' => 1,
     'file:footer.tpl' => 1,
   ),
 ),false)) {
-function content_652b7c0118efc8_24745409 (Smarty_Internal_Template $_smarty_tpl) {
+function content_652ca5337e6dd8_99327200 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_subTemplateRender("file:header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 
 <?php if ((isset($_smarty_tpl->tpl_vars['_SESSION']->value['id_usuario']))) {?>
-    <div>
-        <form method="POST" action="editarLibro/<?php echo $_smarty_tpl->tpl_vars['id']->value;?>
-">
-            <label for="titulo">Titulo:</label>
-            <input required type="text" name="titulo" id="titulo">
-            <label for="autor">Autor:</label>
-            <input required type="text" name="autor" id="autor">
-            <label for="sinopsis">Sinópsis:</label>
-            <input required type="text" name="sinopsis" id="sinopsis">
-            <label for="anio">Año:</label>
-            <input required type="number" name="anio" id="anio">
-            <label for="genero">Género:</label>
-            <select name="genero" id="genero">
-                <?php
-$_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['generos']->value, 'genero');
-$_smarty_tpl->tpl_vars['genero']->do_else = true;
-if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['genero']->value) {
-$_smarty_tpl->tpl_vars['genero']->do_else = false;
-?>
-                    <option value=<?php echo $_smarty_tpl->tpl_vars['genero']->value->id_genero;?>
-><?php echo $_smarty_tpl->tpl_vars['genero']->value->genero;?>
-</option>
-                <?php
-}
-$_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
-            </select>            
-            <label for="precio">Precio:</label>
-            <input required type="number" name="precio" id="precio">
-            <select name="disponibilidad" id="disponibilidad">
-                <option value=0>No disponible</option>
-                <option value=1>Disponible</option>
-            </select>
-            <button type="submit">Editar</button>
-
-            <?php if ($_smarty_tpl->tpl_vars['error']->value != null) {?>
-                <?php $_smarty_tpl->_subTemplateRender("file:error.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
-?>
-            <?php }?>
-
-        </form>
-    </div>
-<?php }?>
+    <?php $_smarty_tpl->_subTemplateRender("file:formularioLibro.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
+}?>
 
 <div>
     <h1>Libros:</h1>
@@ -83,8 +43,9 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
                 <th>Género</th>
                 <th>Precio</th>
                 <th>Disponibilidad</th>
-                <th></th>
-                <th></th>
+                <?php if ((isset($_smarty_tpl->tpl_vars['_SESSION']->value['id_usuario']))) {?> 
+                    <th></th>
+                <?php }?>
             </tr>
         </thead>
         <tbody>
@@ -108,14 +69,16 @@ $_smarty_tpl->tpl_vars['libro']->do_else = false;
                     <td><?php echo $_smarty_tpl->tpl_vars['libro']->value->precio;?>
 </td>
                     <td>
-                        <?php if ($_smarty_tpl->tpl_vars['libro']->value->disponibilidad) {?>
-                            <span class="border border-success">Disponible</span>
-                        <?php } else { ?>
-                            <span class="border border-danger">No Disponible</span>
-                        <?php }?>
+                    <?php if ($_smarty_tpl->tpl_vars['libro']->value->disponibilidad) {?>
+                        <span class="border border-success">Disponible</span>
+                    <?php } else { ?>
+                        <span class="border border-danger">No Disponible</span>
+                    <?php }?>
                     </td>
-                    <td><a href="eliminarLibro/<?php echo $_smarty_tpl->tpl_vars['libro']->value->id;?>
-" type="button">Borrar</a></td>
+                    <?php if ((isset($_smarty_tpl->tpl_vars['_SESSION']->value['id_usuario']))) {?> 
+                        <td><a href='eliminarLibro/<?php echo $_smarty_tpl->tpl_vars['libro']->value->id;?>
+' type="button">Borrar</a></td>
+                    <?php }?>
                 </tr>
             <?php
 }
